refactor: secure all endpoints with auth-based ownership

+++ src/main/java/com/ashutosh/jobApp/service/impl/JobServiceImpl.java
@@ -2,13 +2,21 @@ package com.ashutosh.jobApp.service.impl;
 
 import com.ashutosh.jobApp.entity.Company;
 import com.ashutosh.jobApp.entity.Job;
+import com.ashutosh.jobApp.entity.User;
+import com.ashutosh.jobApp.repository.CompanyRepository;
 import com.ashutosh.jobApp.repository.JobRepository;
+import com.ashutosh.jobApp.repository.UserRepository;
 import com.ashutosh.jobApp.service.CompanyService;
 import com.ashutosh.jobApp.service.JobService;
 import com.ashutosh.jobApp.exception.ResourceNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Isolation;
 import org.springframework.transaction.annotation.Transactional;
@@ -22,11 +30,13 @@ public class JobServiceImpl implements JobService {
     private final JobRepository jobRepository;
     private final CompanyService companyService;
 
+
     @Override
     @Transactional
-    public Job createJob(Job job , Long companyId) {
+    public Job postJob(Job job) {
 
-        Company company = companyService.getCompanyById(companyId);
+        Company company = companyService.getAuthenticatedCompany();
+
         company.addJob(job);
 
         return jobRepository.save(job);
@@ -56,7 +66,13 @@ public class JobServiceImpl implements JobService {
     @Override
     public Job updateJobById(Job updatedjob , Long jobId) {
 
+        Company company = companyService.getAuthenticatedCompany();
+
         Job job = findJobById(jobId);
+
+        if(!job.getCompany().getId().equals(company.getId())) {
+            throw new AccessDeniedException("Access Denied");
+        }
 
         job.setDescription(updatedjob.getDescription());
         job.setLocation(updatedjob.getLocation());
@@ -65,12 +81,20 @@ public class JobServiceImpl implements JobService {
         job.setMaxSalary(updatedjob.getMaxSalary());
 
         return jobRepository.save(job);
+
     }
 
     @Transactional(isolation = Isolation.REPEATABLE_READ)
     @Override
     public void deleteJobById(Long jobId) {
+
         Job job = findJobById(jobId);
+        Company company = companyService.getAuthenticatedCompany();
+
+        if(!job.getCompany().getId().equals(company.getId())) {
+            throw new AccessDeniedException("Access Denied");
+        }
+
         jobRepository.delete(job);
     }
